(TRX): added SR or candidate link to cast vote first step
